Limit to 30v
The PicoPD is limited to 30v outputs. I think it is because of the
a limit in the regulator that provides 3.3v to the microcontroller.

Prevent users from selecting anything over 30v. Mark AVS source to
have a 30v max voltage.  Mark sources >30v a "not detected" to avoid
them being selected.